General: move Loader and Error components folders into Common, adapt references
